added functionality to pay for events

+++ eventica/my_upcoming_events/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
 from django.db import connection
+from django.contrib import messages
+from django.urls import reverse
 
 
 # Create your views here.
@@ -7,14 +9,48 @@ def my_upcoming_events(request):
     user_id = request.session['user_id']
     city = request.session['city'].lower()
     cursor = connection.cursor()
+    cursor.execute(f"""
+                    SELECT event_id
+                    from ticket T
+                    where user_id = {user_id}
+                    """)
+    paid_events_temp = to_dict(cursor)
+    paid_events = []
+    for event in paid_events_temp:
+        paid_events.append(event["event_id"])
 
     if request.method == "POST":
-        event_id = request.POST["event"]
-        cursor.execute(f"""
-                        DELETE FROM joins
-                        where user_id = {user_id} and event_id = {event_id}
-                        """)
-        return redirect(request.META['HTTP_REFERER'])
+        if request.POST["action"] == 'Cancel':
+            event_id = int(request.POST["event"])
+            price = int(request.POST["event_price"])
+            cursor.execute(f"""
+                            DELETE FROM joins
+                            where user_id = {user_id} and event_id = {event_id}
+                            """)
+            if event_id in paid_events and price != 0:
+                cursor.execute(f"""
+                                UPDATE user
+                                SET balance = balance + {price}
+                                WHERE user_id = {user_id}
+                                """)
+                messages.success(request, 'You have successfully canceled your participation and the event fee has been refunded!', extra_tags='bg-success')
+            else:
+                messages.success(request, 'You have successfully canceled your participation!', extra_tags='bg-success')
+            return redirect(reverse("my_upcoming_events"))
+        
+        elif request.POST["action"] == 'Pay':
+            price = request.POST["event"]
+            event_id = request.POST["event_id"]
+            cursor.execute(f"""
+                            UPDATE user
+                            SET balance = balance - {price}
+                            WHERE user_id = {user_id}
+                            """)
+            cursor.execute(f"""
+                            INSERT INTO ticket values(NULL, {event_id}, {user_id})
+                            """)
+            messages.success(request, 'You have successfully paid for your ticket and the event fee has been deducted from your balance!', extra_tags='bg-success')
+            return redirect(reverse("my_upcoming_events"))
 
     search_title = request.GET["title"] if "title" in request.GET and request.GET["title"] else ''
     event_type = request.GET["event_type"] if "event_type" in request.GET and request.GET["event_type"] else ''
@@ -26,15 +62,6 @@ def my_upcoming_events(request):
                     WHERE J.user_id = {user_id} AND E.name LIKE %s AND event_type LIKE %s;
                     """, [q_search_title, q_event_type])
     events = to_dict(cursor)
-    cursor.execute(f"""
-                    SELECT event_id
-                    from ticket T
-                    where user_id = {user_id}
-                    """)
-    paid_events_temp = to_dict(cursor)
-    paid_events = []
-    for event in paid_events_temp:
-        paid_events.append(event["event_id"])
     return render(request, "my_upcoming_events.html", {
         "city": city.upper(),
         "filter_event_type": event_type.lower(),
